Cartographie des imaginaires.
Add a Java "T.A.I" (Test Analogy Unconscious from Carl Gustav Jung) programme to the "imagination map". The programme run with Gephi as Master Server and can be deploid in local to add more computer on the same database. Actually trying to develop the same configuration with a JSON data base in Angular in order to create a webApp. Let me know if you are interrested in the project. Thanks.